Update README.md for container directory
Update README.md in the transformers4recsys/container directory.

The motivation is to add new sections on the installation of the
NVIDIA Container Toolkit in order to provide better documentation
of environment setup to new and existing users, a process which
involves several potentially non-obvious steps.

This commit captures the recent personal experience of the
commit's author in setting up the NVIDIA Container Toolkit
on a new machine.

These subsections added are:
  + Host OS Setup
  + Setup NVIDIA CUDA Drivers
  + Setup Docker
  + Setup NVIDIA Container Toolkit