Remove pinned keras version
During CI we are downloading multiple versions of TF and inspecting each one
trying to resolve dependencies. This seems to be because we are
specifying an exact version of keras, but a range of possible TF versions.

Remove the pin, since it was put in place as a temporary workaround until
TF 2.7 was released in https://github.com/NVIDIA-Merlin/Transformers4Rec/pull/312.