Improve TOC navigation
Previously, the navigation was set to a specific level and
the result was that some pages offered too many headings and
other pages were orphaned--no way to browse to the page other
than clicking a link on an HTML page.

With the change, only the H1 for each page is shown. This provides
a good balance--all pages can be browsed from the TOC and pages
with several H2 and H3s do not cause the TOC to be "jumpy."